Fixed identity evaluation.
Now an entity only needs to start (EventNew) somewhere in the
trueentity's history and be last seen somewhere in the trueentity's
history.

Thus, an entity that appears out of unseen area then disappears into
it (and that's is its entire history) is correct (in terms of
identity), as well as this same trueentity appearing again but being
marked as new.