Improved "intersection" prepared example.
What happens now: The wrong path is built after the first movement. On
the next movement, the labels/paths cannot reach the colored
detections (what's believed to be blue is too far from what really is
blue), so new labels are created.

We may think this is the "right approach," but how is this error
corrected? What kind of error is it?