Many changes related to StepsBetween > 1:
Change (score-by-replaying) to (score-by-catching-up) so that, when
meta-explaining, we don't recreate the progression of epistemic
states, but rather just explain all the rest in one step. The branch
point is modified (something rejected, probably), and explained in one
step, but then the next step explains all the remaining detections (up
to present time).

Updated PlausibiltyAccuracy metric to divide by the StepsBetween
parameter, so that scores > 2 are not possible.

Fixed bugs in the label-coloring code. Added test case.

Fixed bugs in player showing multiple steps.

Added some debug information for issue 4:
https://github.com/joshuaeckroth/retrospect/issues/4