Improved transitive explanation handling. Transitive explanations are considered separately from immediate (first-level) explanations.
As of commit 849d206e61d3b5f77aa47fed8876e85467515cda, transitive
explanations are considered hand-in-hand with first-level
explanations. Also, transitive explanations may be considered
essentials if they are the only way to explain first-level
explanations (rather than the only way to explain the data, which
would never happen because they are second-level explanations).

What we want instead is to consider two distinct sets of explainers:
first-level explainers and second-level (transitive)
explainers. Transitive explainers will be considered explainers of the
data but will be separately considered from first-level
explainers. The two sets will be evaluated in the same way, and the
hyp with the greatest delta (how much its confidence surpasses rivals)
will be chosen; but the first-level and second-level explainers will
not compete as rivals.

Look at the logic in (find-best-multi) in workspaces.clj:

(cond (not (:best best-immediate)) best-transitive
      (not (:best best-transitive)) best-immediate
      (:essential? best-immediate) best-immediate
      (> (:delta best-immediate) (:delta best-transitive)) best-immediate
      :else best-transitive)