Fixed a bug where an invalid mov was offered and another bug regarding mov conflicts.
Specifically, if a prior belief says a color is at some place, then we
can only hypothesize moves that move the color from that place to
another same-colored detection or gray space.

For the conflicts bug, cannot have two colored mov hyps where one ends
at the time the other starts but in which the earlier one does not end
at the same place the latter one begins.