Reverted change that reduced to one sensor hyp per det; back to two hyps (to/from) per det.
Reason is as follows. Suppose we batch the whole problem. We explain
two sensor reports with a movement. We explain two other reports with
another movement. But suppose the end of the first movement and the
beginning of the second movement are actually connected by a third
movement. Yet there is no unexplained report that instigates
hypothesizing/accepting this third movement between them.